requirements.txt
# Core
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0

# Base de datos
psycopg2-binary==2.9.10

# Variables de entorno
python-dotenv==1.0.1

# Cliente HTTP para API externa
requests==2.32.3
python-decouple==3.8

# (Opcional) Para pruebas
pytest==8.3.4
pytest-django==4.9.0